removed fetch data in main, reinstated as a function in fetchData.mjs, deleted another fetch data implementation in api folder
